requirements.txt
discord.py >= 1.7.3, <= 1.8.0
pytest >= 6.2.5, <= 6.3.0
pytest-dotenv >= 0.5.2, <= 0.6.0
python-twitch-client >= 0.7.1, <= 0.8.0
pylint >= 2.11.1, <= 2.12.0